Moved Infix-style extensions into an Infix namespace
