+++ Assets/LeapMotion/Core/Scripts/Utils/Infix.cs.meta
@@ -1,0 +1,13 @@
+fileFormatVersion: 2
+guid: 69526534626aa244d916dc4f775d36ab
+timeCreated: 1517951457
+licenseType: Free
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
